Refresh spec source, enable auto-publish with spec-prod
The last couple of spec updates were made against the **generated** spec:
https://github.com/WICG/ink-enhancement/pull/34
https://github.com/WICG/ink-enhancement/pull/36

The generated spec `index.html` is not the right source though, it should be
the result of generating `index.bs` with Bikeshed:
https://github.com/speced/bikeshed

This update:
1. refreshes the content of `index.bs` to integrate the updates made to
`index.html` (and fix a couple of Bikeshed errors).
2. setups automatic generation and publication of the spec with Bikeshed, as
proposed in https://github.com/WICG/ink-enhancement/pull/32#issuecomment-964857660
3. drops `index.html` from the `main` branch, and make sure it is ignored to
avoid re-adding it later on by mistake.

This would render #32 obsolete and fix #31 (this had been fixed in the source
but not in the generated spec).

Note: once installed locally, generating the spec is a matter of running:

```
bikeshed spec
```